refactor(sa-reg): move data class to module where it's used, don't introduce more layers of logic
